Move Platform into platform tags
