--- crates/platform-host/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "platform-host"
-version = "0.0.1"
-edition = { workspace = true }
-rust-version = { workspace = true }
-homepage = { workspace = true }
-documentation = { workspace = true }
-repository = { workspace = true }
-authors = { workspace = true }
-license = { workspace = true }
-
-[lints]
-workspace = true
-
-[dependencies]
-serde = { workspace = true, features = ["derive"] }
-thiserror = { workspace = true }
