+++ crates/platform-tags/Cargo.toml
@@ -13,7 +13,6 @@ license = { workspace = true }
 workspace = true
 
 [dependencies]
-platform-host = { path = "../platform-host" }
-
 rustc-hash = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
 thiserror = { workspace = true }
