+++ crates/uv/Cargo.toml
@@ -19,7 +19,6 @@ distribution-types = { path = "../distribution-types" }
 install-wheel-rs = { path = "../install-wheel-rs", features = ["clap"], default-features = false }
 pep440_rs = { path = "../pep440-rs" }
 pep508_rs = { path = "../pep508-rs" }
-platform-host = { path = "../platform-host" }
 platform-tags = { path = "../platform-tags" }
 pypi-types = { path = "../pypi-types" }
 requirements-txt = { path = "../requirements-txt", features = ["reqwest"] }
